cargo fmt, bdd_manager module, force heuristic (broken)
