+++ Cargo.toml
@@ -6,6 +6,7 @@ edition = "2018"
 
 [dependencies]
 clap = "3.0.0-beta.2"
+rand = "0.8.3"
 
 [dev-dependencies]
 criterion = "0.3"
