+++ benches/bdd_building.rs
@@ -1,20 +1,20 @@
 use criterion::{criterion_group, criterion_main, Criterion};
 
-use obbdimal::bdd_manager::{BDDManager, InputFormat};
+use obbdimal::bdd::bdd_ds::{Bdd, InputFormat};
 use obbdimal::input::parser::ParserSettings;
 
 fn build_sandwich_bdd() {
     // Read data from a dimacs file.
     let data = std::fs::read_to_string("examples/assets/sandwich.dimacs").unwrap();
     // Create a BDD from input data (interpreted as dimacs cnf).
-    let _mgr = BDDManager::from_format(&data, InputFormat::CNF, ParserSettings::default()).unwrap();
+    let _mgr = Bdd::from_format(&data, InputFormat::CNF, ParserSettings::default()).unwrap();
 }
 
 fn build_berkeley_bdd() {
     // Read data from a dimacs file.
     let data = std::fs::read_to_string("examples/assets/berkeleydb.dimacs").unwrap();
     // Create a BDD from input data (interpreted as dimacs cnf).
-    let _mgr = BDDManager::from_format(&data, InputFormat::CNF, ParserSettings::default()).unwrap();
+    let _mgr = Bdd::from_format(&data, InputFormat::CNF, ParserSettings::default()).unwrap();
 }
 
 fn criterion_benchmark(c: &mut Criterion) {
