+++ src/bdd/mod.rs
@@ -1,2 +1,3 @@
+pub mod bdd_ds;
 pub mod bdd_graph;
 pub mod bdd_manager;
