+++ src/lib.rs
@@ -2,6 +2,6 @@ pub mod bdd;
 pub mod input;
 pub mod variable_ordering;
 
-pub use bdd::bdd_manager;
+pub use bdd::bdd_ds;
 pub use input::boolean_function;
 pub use input::parser;
